water kinda pours when you tip a bottle now
A droplet cube of water falls from the top of the bottle when tipped. Water is removed from the bottle. Droplet can be absorbed by a bottle, the cauldron,  and another droplet. Currently there is a bug where holding the bottle upside down over the cauldron will repeatedly spawn and destroy the droplet and cause an error which also deletes some water. Water quantities are not accurately transferred yet.