Start of buy order sheet
Added a writing space that can be placed after anything you want to buy that has an int for its count and detects if a feather (another thing I added here) is touching it and increases the count by one. Every time the count goes up the text updates accordingly.