Forgot to make small saves, large update read description.
Added getter methods for writing space

Added Blodus Fruit blueprint

Made it so the buy order sheet actually shows the current total for your order. Reduced chance to not update when adding to an order though it is still possible. There is a bug where you can start an overlap event by pushing the feather into the sheet then hold it in and slide over to another writing space and increase it without triggering a new overlap event on the sheet itself making it not update.

Added a sell box that detects if there is an order sheet inside and how many coins are inside and spawns items according to the sheet if there is enough money.

Does not work properly yet. Spawns bottles and mushrooms when only mushrooms are selected. Does not spawn bottles when they are selected. Does not handle coins well, random deletions and improper math.